+++ kidsping/src/main/java/com/kidsworld/kidsping/domain/user/service/UserService.java
@@ -2,28 +2,8 @@ package com.kidsworld.kidsping.domain.user.service;
 
 import com.kidsworld.kidsping.domain.user.dto.request.RegisterRequest;
 import com.kidsworld.kidsping.domain.user.entity.User;
-import com.kidsworld.kidsping.domain.user.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-
-@Service
-@RequiredArgsConstructor
-public class UserService implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    public User save(RegisterRequest registerRequest) {
-        return userRepository.save(registerRequest.toEntity());
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        User user = userRepository.findByEmail(email).orElseThrow(() -> new UsernameNotFoundException("User not found"));
-        return new org.springframework.security.core.userdetails.User(user.getEmail(), user.getPassword(), new ArrayList<>());
-    }
+public interface UserService extends UserDetailsService {
+    User save(RegisterRequest registerRequest);
 }
